Observation Algorithm - CE processing

+++ FSMlib/Model/FSMtypes.h
@@ -80,7 +80,7 @@ static const char* machineTypeNames[] = {
 typedef unsigned int state_t;
 typedef unsigned long input_t;
 typedef unsigned int output_t;
-typedef unsigned int seq_len_t;
+typedef size_t seq_len_t;
 
 typedef list<input_t> sequence_in_t;
 typedef list<output_t> sequence_out_t;
